PListWriter now working

+++ PListA4J/test/com/feistypeach/pojos/Location.java
@@ -6,10 +6,10 @@ import com.feistypeach.plista4j.PListType;
 @PListObject(PListType.StringType)
 public class Location {
 
-	private float latitude;
-	private float longitude;
+	private double latitude;
+	private double longitude;
 	
-	public Location(float latitude, float longitude) { 
+	public Location(double latitude, double longitude) { 
 		this.latitude = latitude;
 		this.longitude = longitude;
 	}
@@ -25,19 +25,19 @@ public class Location {
 		return latitude + ", " + longitude;
 	}
 
-	public float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
